DbAdapter

Description

Implementation of persistent db storage for user defined portfolio information and quote information retrieved from yahoo

Documentation

dbFile :: IO FilePath

The location of the db on disk

connect :: FilePath -> IO Connection

Initialise the db with a minimal schema and return a db connection. The db will be created if it doesn't already exist.

stringify :: Converters a => ([String] -> String) -> [a] -> [String]

Convert a list of Converters instances into a list of strings. Each converter object has its fields converted to string representations. The list of fields is further converted into a single string using the supplied ([String] -> String) function

Types queryable in the db

data Portfolio

Haskell structure that represents the command line output table including yahoo quotes and portfolio position info

Constructors

Portfolio

Fields

prtfsymbol :: String

prtfcurrency :: String

prtfposition :: Double

prtfstrike :: Double

prtfprice :: Double

prtfchange :: Double

prtfpctchange :: Double

prtfvolume :: Double

prtffxp :: Double

data Position

Haskell structure that contains the portfolio position information entered by the user

Constructors

Position

Fields

symbol :: String

currency :: String

position :: Double

strike :: Double

data Fx

Haskell structure that contains the fx to convert yahoo quotes into the symbol currency entered by the user

Constructors

Fx

Fields

fxToCcy :: String

fxFromCcy :: String

fx :: Double

Db updates

insertPosition :: IConnection conn => conn -> Position -> IO Integer

Add a new position to the db

The user specifies all the fields to be inserted in the db.

insertFx :: IConnection conn => conn -> [Fx] -> IO ()

Add the current fx to the db

The required fxs are inferred from the currencies the user has specified of the price they entered a position in, and the currency yahoo claims it returns quotes in. There are known issues with the yahoo currency so the retrieval of the appropriate fx values from yahoo is not implemented yet

populateQuotesTable :: IConnection conn => conn -> [Symbol] -> IO ()

Add yahoo quotes to the db

The set of symbols to retrieve is inferred from the symbols the user has entered in their portfolio.

Db retrieval

fetchSymbols :: IConnection conn => conn -> IO [String]

Retrieve symbols from the db

This is the set of unique symbols specified by the user as part of their portfolio

fetchPositions :: IConnection conn => conn -> IO [Position]

Fetch the user defined positions from the db

fetchFx :: IConnection conn => conn -> IO [Fx]

Fetch fx from the db

This is the set of fxs we require based on the user defined and yahoo currencies in the portfolio.

fetchPortfolio :: IConnection conn => conn -> IO [Portfolio]

Fetch the portfolio from the db

This is the set of positions defined by the user, supplemented with the latest quote data from yahoo