Main

Description

Command line interface

Can be called with any the following options:

-v, --version. Display the current version

-s, --show. Show the portfolio positions. This displays purely the information entered by the user

-u, --update. Update the latest quote data from yahoo and display it along with the associated user defined portfolio position data

-a, --add. Add a new symbol to the portfolio. The user will be prompted for the symbol, number of units, purchase price and currency

Documentation

main :: IO ()