Yahoo

Description

This module provides an interface to make http requests to yahoo finance to obtain quote data on user supplied symbols. The results are streamed back as YahooQuotes

See

http://www.serpentine.com/wreq/tutorial.html for usage of Network.Wreq.get method (wreq in cabal)

https://www.fpcomplete.com/user/snoyberg/library-documentation/conduit-overview for conduit usage

http://hackage.haskell.org/package/cassava-0.1.0.1/docs/Data-Csv.html for usage of Data.Csv package (cassava in cabal)

Documentation

type Symbol = String

Provide a Symbol type for clarity. These should be the stock symbols used by Yahoo

The YahooQuote type

data YahooQuote

The type used to house the quote csv fields we get back from yahoo

YahooQuote fields

symbol :: YahooQuote -> Symbol

name :: YahooQuote -> String

currency :: YahooQuote -> String

quote :: YahooQuote -> Double

change :: YahooQuote -> Double

volume :: YahooQuote -> Int

Populating YahooQuotes

getQuote :: MonadIO m => [Symbol] -> Source m YahooQuote

Make an http request to yahoo for stock symbols and stream the result back. The list of Symbols are passed to the http request, therefore should be recognised by yahoo.

See 'Lifting Operations' in conduit documentation for orientation on how to process the resulting stream.